refactor(#14): SRP 적용을 위한 메서드 분리
